project/requirements.txt
fastapi
sqlalchemy
uvicorn
pydantic
passlib
bcrypt
python-jose
python-multipart